+++ src/components/typeahead-search/http/actionSearchClient.ts
@@ -60,7 +60,8 @@ export function actionSearchClient( getJson: FetchJson = fetchJson ): SearchClie
 		fetchByTitle( query, domain, limit = 10 ): Promise<SearchResponse> {
 			query = query.trim();
 
-			if ( !query ) {
+			// [todo] [IE11] Remove `&& Promise` conditional when native Promise support is assumed.
+			if ( !query && Promise ) {
 				return Promise.resolve( { query, results: [] } );
 			}